adding support for configurable description lenght
getting rid of the 500 char limit, which is not matrix based for topics. it is now possible to define a max char limit for the length of a description. this effects the 'create' route as well as the 'moderate' route as it is possible to add descriptions to contexts as well as to items. this config key is optional so if it is not defined it will fallback to 500 chars.