fix for media upload in gutenberg
there were two bugs which created the problem to use media blocks in the gutenberg editor. on the one hand it was not possible to upload images at all anymore. this was based on the change of parameters of the 'sendImageMessage' function of the matrix-js-sdk which resulted in a infinity loop after trying to save the changes.
on the other hand an additional problem bricked all valid media uploads. they were still uploaded but not possible to display neither in element nor in the cms. this problem occured with the change of the return of the 'uploadContent' function of the matrix-js-sdk, which now returns an object with the key 'content_uri' instead of a string. a bugfix for this problem was implemented in the main branch before already for the thumbnail upload.